+++ Migrations/202101130225508_produtoApagadoNew.cs
@@ -1,0 +1,18 @@
+namespace TP_PWEB.Migrations
+{
+    using System;
+    using System.Data.Entity.Migrations;
+    
+    public partial class produtoApagadoNew : DbMigration
+    {
+        public override void Up()
+        {
+            AddColumn("dbo.Produtos", "Apagado", c => c.Boolean(nullable: false));
+        }
+        
+        public override void Down()
+        {
+            DropColumn("dbo.Produtos", "Apagado");
+        }
+    }
+}