+++ Models/Promocao.cs
@@ -11,11 +11,16 @@ namespace TP_PWEB.Models
     {
         [Key]
         public int IdPromocao { get; set; }
-       
+
+        [Display(Name = "Produto")]
         public int IdProduto { get; set; }
         public virtual Produto Produto { get; set; }
 
         [Display(Name = "Promocao (%)")]
         public int Percentagem { get; set; }
+
+        public decimal PrecoNovo { get; set; }
+
+        public bool Ativa { get; set; }
     }
 }